requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiohttp-socks==0.6.0
aiohttp==3.7.4.post0
apscheduler==3.7.0
asgiref==3.3.4; python_version >= '3.6'
async-timeout==3.0.1
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
certifi==2020.12.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django-apscheduler==0.5.2
django-cors-headers==3.7.0
django==3.1.7
djangorestframework-simplejwt==4.6.0
djangorestframework==3.12.4
easydict==1.9
filelock==3.0.12
hiredis==2.0.0; python_version >= '3.6'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
joblib==1.0.1; python_version >= '3.6'
lxml==4.6.3
ml2rt==0.2.0
multidict==5.1.0; python_version >= '3.6'
numpy==1.20.2; python_version >= '3.7'
packaging==20.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
protobuf==3.16.0
psycopg2==2.8.6
pyjwt==2.1.0; python_version >= '3.6'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-socks[asyncio]==1.2.4
pytz==2021.1
pyyaml==5.4.1
redis==3.5.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
redisai==1.0.2
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
scipy==1.6.3
sentencepiece==0.1.91
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
tokenizers==0.9.3
torch==1.7.1
tqdm==4.60.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
transformers==3.5.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
yarl==1.6.3; python_version >= '3.6'
